Classes vs Dicts
Dictionaries are fine and all, but over use and nesting can result in
some crazy code that is both hard to read and hard to work with.

Try implementing a class structure with the smallest bits of data in
dicts or names lists. This allows for far more readability, scalability,
and flexability