Interfaces & Hooks
When providing a function as a hook to a method call, consider using a
class container to track stateful behaviour. This works due to python's
first class function implementation, allowing functions to be passed
around like variables. Additionally, consider implemnting __call__ so
that the class instance can be called like a class first function
itself.