Mixins & Multi-Inheritance
Using mixins is one of the few good examples of multi-inheritance.
They allow for simple encapuslation of behavior into various different
classes that can be inheritated anywher along a class heriarchy,
facilitaing the addition of complex bahviors in a clear and precise
manner. As always, feel free to override method behavior to enhance or
work around bugs for any classes that inherit the mixins.